app/requirements.txt
python-dotenv
pandas
mecab
Mastodon.py